requirements.txt
setuptools==67.0.0

pytorch-lightning==2.2.5

# huggingface_hub
huggingface-hub>=0.22.0 
diffusers >= 0.9.0
accelerate
transformers

supervision==0.17.0
kornia
matplotlib==3.5.2
numpy==1.26.4
plyfile==0.8.1
open3d 


opencv-python==4.7.0.68
opencv-contrib-python==4.7.0.68

imageio==2.25.0
imageio-ffmpeg 
scipy==1.15.3
scikit-image==0.19.0
black 
chumpy 
rtree 
tensorboardX 
tensorboard

transforms3d 
munch
ipython 
jupyter 
pycocotools

rich==13.5.3
moviepy 
pyrender==0.1.45 
shapely==1.8.4 
trimesh==3.12.6 
triangle
pydot

omegaconf

numba
hydra-core
openpyxl

cython

ninja
PyOpenGL 
glfw 
xatlas 
gdown

timm==0.9.10

configargparse
easydict

Pillow==9.5.0

# marigold
accelerate>=0.22.0
diffusers>=0.20.1
transformers>=4.32.1

einops
wandb

# tapier
ipympl 
mediapy==1.2.2
# chex
einshape

cupy-cuda12x==12.2.0

pillow
imageio
joblib
pandas
scikit-image
scikit-learn
scipy
seaborn
PyOpenGL-accelerate
pyrender
ninja

tqdm
munch
evo
open3d==0.17.0
torchmetrics
imgviz
lpips
rich
kornia
PyQt5
glfw
PyGLM
mmengine

git+https://github.com/deepmind/einshape
jax==0.4.13
https://storage.googleapis.com/jax-releases/cuda11/jaxlib-0.4.13+cuda11.cudnn86-cp310-cp310-manylinux2014_x86_64.whl
flax==0.7.1
jaxline==0.0.5
ml_collections 
opencv-python==4.7.0.68
optax==0.1.4
chex==0.1.82
dm-haiku==0.0.10
dm-tree==0.1.8
typing_extensions==4.7.1